Factor out theme into a separate module
